requirements.txt
torch
transformers
datasets
peft
bitsandbytes
accelerate
trl
wandb
fastapi
uvicorn
gitpython
python-dotenv
pyyaml
requests
